Enhance CarPLAN model and data processing: - Update TransformSamplingSubTraj to include reference lines in the transformation. - Modify CarPLAN class to conditionally initialize displacement encoder and predictor based on is_carplan flag. - Refactor training step to incorporate reference lines for improved planning loss calculation. - Adjust PointsEncoder to ensure proper tensor type handling. - Clean up unused code in MapEncoder and PlanningDecoder. - Add set_path.sh script to set up the repository environment.

+++ src/policy/latentdriver/layers/fourier_embedding.py
@@ -50,6 +50,6 @@ class FourierEmbedding(nn.Module):
         x = torch.cat([x.cos(), x.sin(), continuous_inputs.unsqueeze(-1)], dim=-1)
         continuous_embs: List[Optional[torch.Tensor]] = [None] * self.input_dim
         for i in range(self.input_dim):
-            continuous_embs[i] = self.mlps[i](x[..., i, :])
+            continuous_embs[i] = self.mlps[i](x[..., i, :].to(torch.float32))
         x = torch.stack(continuous_embs).sum(dim=0)
         return self.to_out(x)
